# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/medicos": {
            "get": {
                "tags": [
                    "Medicos"
                ],
                "summary": "Listar todos os médicos",
                "operationId": "c7cebce3506c331d127872eee61c4ded",
                "responses": {
                    "204": {
                        "description": "Lista de médicos retornada com sucesso"
                    },
                    "404": {
                        "description": "Nenhum médico encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/medicos/{id}": {
            "put": {
                "tags": [
                    "Medicos"
                ],
                "summary": "Atualizar um médico",
                "operationId": "0263d24c7869aded8d123a8f045ab68d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf",
                                    "crm",
                                    "especialidade_id",
                                    "status"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Dr. Rodrigo"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "crm": {
                                        "type": "string",
                                        "example": "CRM123456"
                                    },
                                    "especialidade_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Médico atualizado com sucesso"
                    },
                    "404": {
                        "description": "Médico não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Medicos"
                ],
                "summary": "Criar um novo Medico",
                "operationId": "00493663b2ef311c609ebb0ae89ebad5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "cpf",
                                    "crm",
                                    "especialidade_id",
                                    "status"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Dr. Rodrigo"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    },
                                    "crm": {
                                        "type": "string",
                                        "example": "CRM123456"
                                    },
                                    "especialidade_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Medico criado com sucesso"
                    },
                    "404": {
                        "description": "Erro ao criar Medico"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Medicos"
                ],
                "summary": "Remover um médico",
                "operationId": "ecda6035bb1152f7f6297e1ab836889c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Médico removido com sucesso"
                    },
                    "404": {
                        "description": "Médico não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}